Use SCIM format in event publisher
We want to use the SCIM format for messages. Currently a draft standard
at <URL:https://tools.ietf.org/html/draft-hunt-idevent-scim-00>. We are
not completely in sync with the attribute names. Also, the API merges
messages from 'different' events in Cerebrum, so that e.g. all events
related to a creation is sent as one message, and not create followed by
setting of name, etc.
Jira: CRB-1612