Exchange-integration recovers when session busy
Sometimes, a command sent to Exchange over the
WINRM/PSSession-transport, decides to wait for input or crash. This
results in the PSSession assuming a busy-state, where it is
unrecoverable.

If this occurs, we attempt to detect the busy state, and kill the
PSSession. If that succeeds, a new PSSesssion should be created the next
time around.