Remove hard coded ssl.PROTOCOL_* constants
This commit removes the hard coded ssl protocol constants in
Cerebrum.https, and moves the SSLConfig constants to a new SSL_VERSION
map.

The map is ordered by version value (lowest version to highest version),
so it is looks prettier when the keys are printed. Also, it makes it
easy to fetch the lowest or highest available version (first and last
value in the iterator).

The new SSL_VERSION map is used in the bofhd server script. This now
makes it possible to start bofhd in python environments where a
certain `ssl.PROTOCOL_*` constant was missing.