Add 'description' to Account objects
Adds the 'description' attribute back into account objects. The database
schema already has support for this value, but access to it through
Account was never implemented, or removed at some stage.

NTNU has this functionality, so in order to try and synchronize the code
bases, we add back:

* an `Account.description` writable attribute
* lookup of description in `Account.find()`
* writeback of description in `Account.write_db()`
* setting an initial description in `Account.populate()` and
  subclasses (except the virthome account classes and PosixUser classes)
* setting an initial description in `Account.new()`. No subclasses
  overrides this.

Issue: CRBPRE-37