Remove `Account.create()`
Removes the `Account.create()` method. It was only used by one script
(the probably unused contrib/generate_accounts.py).

The method popoulated a new Account object, added spreads from
BOFHD_NEW_USER_SPREADS, and generated a password, while making sure
everything was written back to the db.

It's a good idea to have one or multiple template methods to create
accounts, so that we don't need to repeat that procedure in multiple
locations, but `Account.create()` is not the correct way to implement
this.

Ideally, we should have an `AccountCreator` class, or something similar,
with features for multiple account types, converting accounts between
account types, migration of existing accounts when new procedures for
creating accounts are added, etc...

Issue: CRBPRE-37