Add possibility to use non-cerelog handlers
Adds code from NTNU, based on logging.config, that makes it possible to
use classes from `logging.handlers` (given as `handlers.<name>`), and
sub-packages (given as `<module>.<class>`) from a logging config dict or
file.

This is a temporary fix, to make our code compatible with the NTNU fork.
The entire cerelog module is due for a rewrite to get rid of the 'one
global logger' practice.

Issue: CRBPRE-56