Change treshold for FS changes to 50%
This is to avoid getting false positives at the start of each semester,
thus stopping the student access process when it is at its most hectic.

By setting the same threshold everywhere we simplify the import a little
bit. Since we want to start using FSWS in the near future, I make it
easy and don't implement cereconf values.

Resolves: CRB-2246 (a bit too simple than original requested)