Add a new logging module
Add a new module `Cerebrum.logutils`, which should replace
`Cerebrum.modules.cerelog`.

The new module implements the functionality of `cerelog`, but in a more
flexible manner. The main reason for this, is to allow a more flexible
configuration of the individual loggers. This is something that
`cerelog` does not facilitate. With this ability, we can move towards
using module loggers (`logging.getLogger(__name__)` in a module), rather
than passing around a global logging object, or relying on
`Cerberum.Utils.Factory.get_logger` to return the same object on each
call.

Options
    Options in this new module are processed by `argparse`, rather than
    processing `sys.argv` with a custom parser. This allow us to
    integrate logger options with the existing parsers for scripts, and
    get help texts for the logging options.

Module config
    The module is configured by `Cerebrum.config` configuration objects,
    which reads config files from a default location. This module will
    look for a config named `logging` and apply it.

    This config determines the base directory for logs, and the
    behaviour of exceptions and warnings (should they be logged,
    etc...).

Logger config
    The module configures the logger hierarchy with files using
    `logging.config.fileConfig` and `logging.config.dictConfig`.

    It finds and loads a config file that matches the `logger-name`
    (option `--logger-name`, or `getLogger(logger_name)`.

    ini-files are loaded using `logging.config.fileConfig` without
    disabling previous loggers, and all other config files are parsed
    using `Cerebrum.config.parsers` and loaded using
    `logging.config.dictConfig`.

Issue: CRBPRE-63