Add a /person/<id>/accounts name attribute
The account name is available under the 'id' attribute. This is wrong,
and the attribute *should* contain the account entity_id, if anything.

This commit adds a 'name' attribute, but keeps the 'id' attribute in a
transition period, until we can make sure that all systems that use this
attribute has been updated.

The 'id' attribute should be removed, or include the entity_id of the
account in question.

Issue: CRB-2330