Change logger preset directory config to be a list
The logger 'confdir' has been renamed to 'presets', to avoid confusion.
We now have a logging environment config, and a logger *preset* config.

The presets setting (previously confdir) is changed from being a single
directory to look for presets in, to a list of directories to look for
presets in.

This allows us to e.g. have a directory with *common* presets, and a
directory with *environment-specific* presets.

Issue: CRB-2387