Fix atomicfile arguments
Fixes Cerebrum.modules.atomicfile to have a signature similar to
io.open. The change in argument order (`replace_equal`), should be
compatible with other Cerebrum code (as the default argument is not
actually modified anywhere).  Also, the default values for `errors` and
`encoding` are set to the same as `io.open`.

Also, the 'a' mode now use shutil to copy the original file for append
mode. This simplifies some of the ugly file mode logic related to
append-mode.

Issue: CRB-2414