Stop updating name for persons not employed
When a cleaning job runs late at night, names are removed from persons
who don't have any affiliations from the same source system as the name.

The addition and removal of a persons name triggers recalculation of a
persons email-address.

This address is then distributed back to the HR-system and stored. The
storage results in a message describing the storage being sent, which in
turn triggered importing the persons name from the HR-system to
Cerebrum.

If the person has different names in the HR-system and some other
authoritative system, the above mentioned import would re-trigger
email-address recalculation, and export the new address to the
HR-system.

This results in infinite email-address oscillation for the affected person.

The fix is simple. Don't import the name.