Fix an issue with ou names in the ABC import
This fixes an age-old problem where the ABC import randomly fails
because a required OU name field (dict key) is deleted from the
interrim OU object.

The fields gets deleted when the field name is *not* translated (or
translated to the same literal value) according to abcconf.CONSTANTS,
because an ``is`` test is used for comparison. The *is* test may fail if
the string literal is *not* the same object (which happens more often
with unicode objects on Python 2.7).

This commit removes the test entirely, and re-builds the entire field
dict.

Issue: CRB-2533