Fix consent update and changlog
The consent module does not update correctly.  If we ever were to update
a consent description or expire date, *all* consents would be affected.

In addition, the consent module logs changes prematurely.  Changes are
logged when adding, removing, or updating consents in the internal
consent-mixin cache.  These changes are not written to the database
until ``write_db`` is called, but the *changes* are logged immediately.

This commit extracts all sql and db communication into separate
functions.  This makes it easier to catch similar errors, as well as
making it possible to perform consent-changes without interacting with
the consent-mixin.

Issue: CRB-2544