List the last changes in history commands
When optimizing entity history, I neglected to observe that the user is
interested (in most cases) in the last changes made to an entity. The
LIMIT ensured that we would get only the oldest changes.

We could have fixes this by ordering DESC and then reversing the list of
the result, but that solution implies a redesign in order to reduce
complexity and binding.

Revert "Perform LIMIT for BOFHDs history commands in DB"
This reverts commit 5c9b5880fa2b477fa5415e06137511b5cf51a849.

Resolves: CRB-2554