Fix an issue where change_log formatting fails
When formatting the 'change_by' field in a change_log record, if both
the `change_by` and `change_program` is `None`, formatting fails. The
failure occurs because the formatting tries to look up the entity name
of `entity_id=None`, and the lookup doesn't handle that.

The scenario *shouldn't* occur, but it did. This fix improves the error
handling, so that the name appears as 'notfound:None' in those cases.

Issue: CRB-2577