Rework logging in multiprocessing
This changes the log record serializing, so that it handles
unserializable data and incorrect log message formatting better.

In addition it adds a max size on the log record queue, so that it
can't fill up and consume too much memory. A separate thread monitors
the log queue size and warns if the queue gets too big. Hopefully this
will halt the logging process rather than kill the queue manager.

Issue: CRB-2572