Fix issues with pip install/uninstall
Cerebrum uses the `data_files` parameter to create a directory
structure, and potentially install some default config files on install.
The default configs are only installed if no modified config file exist
on the target location.

When using pip to install packages, pip will record which files are
installed for any future uninstall. This includes files and directories
from the `data_files`-argument to `setup()` from distutils/setuptools.

With older versions of pip, if a directory was specified as a
`data_files` without any actual files, the directory would be created
with the specified permissions. However, the directory name was also
recorded and queued for uninstall.  pip would then remove these
directories when running an upgrade or unintsall, even if the directory
was no longer empty.

This is fixed in newer versions of pip -- pip never creates empty
directories from `data_files`.

This means that:

1. With older versions of pip, we risk losing log files and config files
   if Cerebrum is upgraded or uninstalled.
2. With newer versions of pip, we can't actually create the directory
   structure we want using `data_files`.

To solve this issue, and ensure that no install/upgrade issues will
re-occur, we remove all hacks related to creating empty directories and
installing default configs.

Issue: CRB-2658