Fix an issue where bofhd blocks new connections
When calling SSLSocket.accept() on a listening ssl socket (in blocking
mode), the ssl module will ensure that an ssl handshake is performed,
and that the client socket is wrapped in an SSLSocket().  This also
means that SSLSocket.accept() blocks and waits for the client to
initiate handshake on the socket.

As the bofhd server waits for this to complete, no other connections can
be accepted.  If a client connects and never initiates an SSL handshake,
the bofhd server process will prevent (or really, queue) all other new
connections.  This means that a malicious (or a buggy) client easily can
prevent other clients from connecting to bofhd.

This commit adds a timeout that prevents the `accept()` call from taking
too long, using `signal.alarm()`.

A better solution *may* be to implement the SSLServer using a
non-blocking SSLSocket, and handle the ssl handshake in the client
thread.

Issue: CRB-2699