Delete the password of a user being deactivated
Remove the row containing information about the user from
account_authentication in the database.
This should be enough to fix the issue but:

Most of proc_delete_user in process_bofhd_requests should probably
be using Account.deactivate(). To do this Account.deactivate() must
be rewritten.

It does not look like Account.deactivate() is really used for anything
smart elsewhere, but not completely sure. If it is it would take time to
rewrite it as the things that depend on it also might need to be rewritten.

Issue: CRB-2686