Check if None is passed to find inside entity.find()
Just test if entity_id is None, and if it is raise a notfounderror
This is easier than going into every function that runs entity.find()
and make them test for None themself.

Could make another error when None is passed. But I believe it
makes sense to say that you couldnt find anything.

Issue: CRB-2696