Stop trying to demote posix groups that can't be demoted
A posix group that is the file group for a posix user can't be
demoted because the user needs it.
Deleting would violate foreign key constraint in the database.

Add a method to PosixGroup to check if it is a file group for a user.
Use this method to check for problem in group_demote_posix command before
running demote_posix. Also used to return False when running demote_posix in
PosixGroup if the demote can't be done.

Issue: CRB-2622