Make it possible to only get accounts with disabled quarantines
There is a need to get accounts that have disabled quarantines in addition to
those that are active. The only_active that exists allows you to either pick out
all quarantines wheter they are active or not. Or to pick out just those that
are active.

Simply pick out the entities that are disabled by checking if they have a
disable_until date set, and if that date is in the future.

Issue: CRB-2639