Stop mailing accounts that have a disabled password quarantine
If an account is quarantined because of old password and the quarantine is then
disabled after without changing the password. The account will currently be seen
as an active account with an old password and then the quarantine process and
sending notifications about the need to change password will start again.

This is unnecessary as the quarantine has been disabled, so the user should know
that the password is old. If someone had the ability to disable the quarantine
they don't need to be reminded that their password still is old.

Issue: CRB-2639