Fix host_a_remove for hosts with IPv4 and IPv6
The bofhd command host_a_remove would not allow users to delete the last
AAAA record, for hosts with existing A records if the host also had
CNAME or SRV records.
The inverse is also true (unable to delete the last A record despite the
host having AAAA records)

The fix is to check if the host has any A *or* AAAA records (regardless
of which IP type is removed) before checking if the host has remaining
CNAME or SRV records.