Fix shell value in posix exports.
This commit fixes an issue where the shell is set to the ShellCode
string value, and not the ShellCode description.

When exporting shells for a user, the shell should be set to an absolute
path to the relevant shell. That value is stored in the ShellCode
description, and not the actual constant/strval.

Issue: CRB-1037