Clean up use of Stedkode class
There was a lot of occurances where we imported the Stedkode class
directly instead of using the Factory.get('OU')(db) method. This has now
been fixed, and instead of importing UiTs StedkodeMixin, that class is
now added to cereconf.CLASS_OU so that the Factory.get('OU') method
imports it properly.

We also stop using the sitoFactory method as the regular Factory method
does the job fine by using the regular OU object when the UiT specific
mixin is included in cereconf.CLASS_OU

Issue: CRB-2756