Add a EntityName.entity_name_exists method
Adds a method to check for the existance of a given entity name,
optionally filtered by value domain(s).  This method should be more
effective than doing `EntityName.find_by_name()`, as it doesn't populate
a Entity object.

This method is introduced to eliminate a query in the UiT account mixin.
It should also be able to replace a lot of `find_by_name()` calls that
simply checks for the existance of a name.

Issue: CRB-2776