Add a try except clause for group name lookup
When running bofh command 'user history username' the operator would get
an Error if the changelog of user contained a reference to a group that
no longer existed.

This is possible since we do not completely delete entities, but rather
remove their names. Thus one can find that there is an entity with a
certain entity_id, and that it has the entity_type group, but when
tyring to look up at group the group does not exist as it has been
deleted.

Issue: CRB-2784