Refactor UiTs OrgLDIF
Changes the init_accounts_info method to call the super and then remove
the sito accounts afterwards. This way we avoid having to update code
two places if we want to add anything extra to the account_info method.
The old way should be slightly more efficient but the performance hit is
worth it to make the code base easier to maintain.

Also removed UiO remains that was never used.

Issue: CRB-2749