Remove uit import_OU.py and event_publisher hack
This commit removes no/uit/import_OU.py, as the script is nearly
identical to no/uio/import_OU.py.

UiT had a patched copy of Cerebrum.modules.event_publisher.event to
deal with a bug in their copy of import_FS.py, this hack is no longer
needed, and is removed as well.

Issue: CRB-2514