Stop importing the AD module in powershell
Windows Server 2019 spews a warning when importing the AD cmdlets::

    WARNING: Error initializing default drive: 'Unable to contact the server. This may be because this server does not
    exist, it is currently down, or it does not have the Active Directory Web Services running.'.

This warning is included in the stdout that we get. Parsing JSON gets
hard.

Running `Import-Module ActiveDirectory` is not needed, though, since
Powershell 3 (included in Windows Server 2012 [1]) supported
auto-loading of such modules [2]. Simply not importing the module
removes the warning, while the cmdlets execute seemingly as normal.

This makes us not support versions lower than Windows Server 2012, which
are out of support anyway.

Only tested in two different environments, for now.

1. https://en.wikipedia.org/wiki/PowerShell#PowerShell_3.0
2. https://docs.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_modules