Fix an issue with no/uit/process_students.py
Fixes an issue where the uit version of process_students always updates
account expire_date, regardless of the current value.

We now only update the expire_date if a change is needed, i.e.:

- If the person is deceased and the current expire_date is not the
deceased date
- If the current expire_date is not the default_expire_date for this run

Also, we only update the expire date for spreads if the expire date for
the user needs an update. This should reduce the running time for this
script.