Reduce log level of log message error -> warning
There are instances where the target of a person role is not specified
well enough for us to process it. This does not necessarily mean that
there is a problem that anyone needs to do anything with, but we still
want to log it such that we can do something with it if there is a
problem.