Add job runner client script.
Adds a client-only version of servers/job_runner/job_runner.py. This
script does not have the unfortunate side effect of starting a new
job_runner daemon if no arguments are given.

The script is installed to <sys.prefix>/bin/, which should ensure that
it is available on the PATH in most environments.