Optimize NISUtils
We now gather all the members of the top level in one chunk instead of
looping through every single group and finding its members. This speeds
up the process considerably on the dev environment. Hopefully that is
also the case for the production environment.

Previously the members of a group were added to the list of members of
that group's parent group. This meant that a group with a single group
as member got all of the members of that sub group as members in the
output file. In additition the sub group also got the members meaning
that when imported there would be double entries for a lot of users.
This no longer happens.

There is still some optimization that can be done by doing a collective
db query for all members below the top level as well but that involves
keeping track of which groups the subgroups are members of and which
groups those groups are subgroups of and so on, which may end up being
more overhead than it's worth. Still worth checking out if we need this
to go faster.

Issue: CRB-2828