Add a job_runner summary report script
Generates a reStructuredText(rst) file with information on each job
defined in scheduled_jobs.py.

The extra bit we get from this compared to just dumping the jobs from
job_runner is the docstring of the scripts the jobs are running.

This does not include a solution for non python scripts, and python
scipts without docstrings. Furthermore, jobs that run the same script
with different arguments still get the same docstring.

Issue: CRB-2870