Clean up import_SAP_person noise
Before this, import_SAP_person.py would try to delete the entity address
of a person if the data in the sap import file was erroneous. This
resulted in a new change log entry every time the source data was wrong,
filling up the entity history with these events.

This commit adds a check to see if there is an address of the
corresponding type to be removed from a person before we delete it.

Did some clean up of the code in general as well. It's now PEP8
compliant.

Issue: CRB-2771