Raise exception if powershell command too long
Due to the limitation on the command line length for powershell command
we ran into trouble creating groups with a large amount of users. This
should have been picked up by a try except block, but powershell does
not always return appropriate errors, making it fail.

This adds a check and raises the appropriate error before sending the
command, making it possible to catch the exception and deal with it.

Issue: CRB-2804