Remove users without affs from groups
Adds a consumer that checks if a person has lost all its affiliations
when an affiliation is removed. If that is the case all the accounts
connected to the person are removed from all groups that are not the
account's personal file group.

Add function for removal of group members

This new function allows removal of group members from a group without
populating the group object. Useful when you know the group id and that
the user is in fact in the group.

Issue: CRB-1886