Update numval when in reminder mode
The job said it would update the numval when in reminder mode. But because the
scheduled job sending out reminder messages didn't have the min_attempts
variable set it didn't actually do it.

Made it so that when a message is actually sent the trait is updated.

This solution is not supernice since the log message might not make as much
sense as it hopefully could have done. Better solution would probably be to
seperate the welcome sms part and the reminder sms part as they are different
enough to deserve that.

Will also need a one time script to set a numval for all old sms_welcome traits.
Because currently the logs are spammed with messages of the script trying to
delete the welcome_sms trait but failing because reminder won't remove traits.

Issue: CRB-2877