Fix delete_spread override method
Entity.delete_spread only takes one argument. (not counting the instance object)
AccountEmailMixin has the possibility to pass in multiple, which will lead to an
error if it is ever done.
PosixUser will always pass multiple because the instance object was passed as an
argument explicitly in addition the actual argument(s)

Issue: CRB-2842