Fix bug in NISUtils
Didn't catche group names correctly resulting in looking up the name of
a group with no name cached, returning None. Whcih we then did a "
".join() on, making the job crash. This is now fixed, and all
group names are cached when dealing with groups.

Modifed the list_names method of EntityName to allow a list of
namespaces to save some time on a SQL query.

Issue: CRB-2828