Stop send_welcome_sms from spaming the log
In reminder mode send_welcome_sms will go over many rows it is not supposed to,
but when it tries to remove the trait nothing happens as we want to keep the
trait.

This meant that the next day when the script ran again it would again try to go
over every row and log a message that it tried to remove it but could not
because it was in reminder mode.

The scipt now updates the numval whenever it tries to remove if it is in
reminder mode. This stops it from going over the row again later.

Issue: CRB-2877