Remove error when attr in const but not clconst
If an attribute was found in const but not in clconst having clconst lookup as a
default in getattr threw an AttributeError since it could not find it. Even
though everything was supposed to have been fine since it was found in const, so
it shouldn't need to go to default.

Fixed by making it only look in clconst if it can't find the attribute in const.

Also switched around the return on change_by because operator never is None when
picking out rows from the database, meaning that it would always default to the
operator even when a program existed. Program is most interesting if it exists.

Issue: CRB-2574