Reduce Constants transaction error log level
Reduce the log level of the error message given when the global
_CerebrumCode transaction is broken and needs to be created.

This happens regularly in threaded applications; if multiple threads
does a constants lookup at the same time, one of the transactions gets
rolled back, and the other transaction savepoint is invalid.

We do however have code to re-create the transaction, which is read-only
anyway, so it doesn't really cause any problems.