Remove expire_date values from Group.search_member()
This commit removes the expire_date from the search_member() results.

1. For ordinary groups, search_member() has to mutate the result in
   order to get a single expire_date value.
2. That expire_date value may be the group expire_date or the member
   expire_date, no way to know for certain. This makes the value
   somewhat unusable.
3. The expire_date is really only used by the group_list bofhd command,
   which has to look up the member to fetch other values. We, may as
   well fetch the expire date that way.

All in all, removing the expire_date means we get to remove complexity
that is not really used, and shoudn'd be used.  Also, since we don't
need to mutate the database row objects, we're one step closer to
replacing db_row.