Clean up change params for audit log
Registers various translations of parameters for different changelog
events. Instead of logging the int values of various constants we now
log the str values instead. This is useful in cases where the constants
no longer exist for some reason.

Makes two minor changes to dns so that changes are logged now.

Issue: CRB-2589