Rename ChangeLogConstants
Rename all constants to the new agreed upon standard of ('nouns',
'verb'), where the verb is one of
add/remove/create/delete/set/clear/update/modify and the nouns are
whatever we do this to.

The change is also reflected in all the mappings of events to these
constants.

Issue: CRB-2896