Remove printer quota info from join_persons
The only thing left of printer quotas is the flag for whether or not a
person is exempt from the quota regime. Instead of putting the logic for
how this is chosen inside the join_persons script we instead leave the
job of setting this to the quota update script.

Issue: CRB-984