Add and remove functionality from Uit FS import
UiT has been running UiO's version of the script. This adds the missing
functionality to UiT's own script so that they can use their own one
instead.

In addition we remove all mention of printer quotas from the UiT import
since the imported printer quota information is not used by anything in
cerebrum.

This includes a commit in the config repository where we switch the
script in scheduled_jobs.py

Issue: CRB-984