Coerce db_row results to dict
By some strange reason, the db_row object returned by does not like
dict-like lookups even though the fields are present in the object.
Coerce the object to a dict before solves it, but still feels a litle
hacky given that db_row should allow for such lookups in the first
place.