Add new person affiliation status
Adds the person affs status TILKNYTTET/komitemedlem that is meant to be
used for the new users in the hiring committee that need access to
JobbNorge.

We achieve this by giving them a limited UiO account, populating this
to LDAP, which gives the users access to the WebLogin service. This
can then be used to login to JobbNorge.

Issue: CRB-2724