Find groups and accounts at the same time
The sync is taking a lot of time. Almost all of that time is used to search for
members in groups. By searching for groups and accounts that are members at the
same time the amount of queries too the database is almost halved.

Took out the building of the members dictionary to its own function to make the
code more easily readable

Issue: CRB-2933