Fix bugs in ou_disk_mapping module
A status constant was not checked if valid or not in the aff_lookup
function. This resulted in the users being able to write the names of
wrong constants and getting a server error instead of the more
appropriate "Unknown affiliation" message.

The get and delete methods of the OUDiskMapping class could get and
delete multiple rows. This has been fixed and they can now only remove
single rows.

Issue: CRB-2719