Change how ou groups work
With this change, ou groups which belong to ous that are
quarantined will show up as empty. This has been done so that recursive
ou groups will behave like the automatic groups which are generated with
populate-automatic-groups.py. Then it is possible to compare the members
of ou groups and automatic groups, to verify that the members of the
newly generated automatic groups are correct.

Issue: CRB-2933