Stop lowering external id when searching
When searching for external ids we used LOWER() in the sql query, but if
the locale of the database is not set, certain characters are not
lowered. This made a search for an external id with the character Ø in
it fail to show any results when there was an entry in the database.

This commit removes the LOWER from the sql query, making the search
case-sensitive. We also stop lowering the related value we search for so
that we actually find the entry we are looking for.

Issue: CRB-2970