Fix issue with bofhd command misc password_issues
The command is used to see if a user can use the new sms reset password
service. This fixes an issue where the command would wrongly indicate
that a user could not use the password service. If an account is
considered to be "fresh" it can in fact use the password service even
if a new phone was added recently.

Issue: CRB-2921