Change how spread_cleaner works
Instead of caching the spreads to keep, we now search directly for
spreads to delete. The approach where a keep_cache is created first may
have caused problems when running spread_cleaner simultaneously with
other jobs.

This script used to call the method notify_spread_expire on each
non-expired spread, which was very inefficient and created a lot of
noise in the logs. That has now been changed so that a lookup is done
in the spread_expire_notification table first.

Issue: CRB-2910