Set maxsize for event publisher queue
Before this commit the queue was allowed to grow infinitely. Because we
suddenly published a large amount of events the workers were not quick
enough to publish the events from the queue, and the queue grew large
enough to fill the memory of the production environment. This made the
machine sad, making it kill the process.

With the max size set for the queue the collector waits for the queue to
have more room before putting anything else on it.

Also swapped from the multiprocessing.Queue to the Queue.Queue

Issue: CRB-2994