Add a new check to the remove_expired_groups script
Previously we did not check if an expired group owned any accounts. This
meant that we tried to remove an expired group which failed due to
a database constraint since it owned accounts. This is not something
that should ever be the case, but you know.. ¯\_(ツ)_/¯

This commit adds a check for this and logs a warning if it occurs.

Issue: CRB-2892