Allow removal of contact info from source systems
This commit removes a bofhd command contstraint that disallows removing
contact info from a given source system if the owner also has an active
affiliation from that source systems.

This is a quickfix so that ops are able to remove contact info values
that disappears from the source system, and aren't automatically removed
by the integration.

It would probably be better to move/re-implement the source system check
to `can_remove_contact_info`, so that e.g. only superusers can remove
values that are set from source systems?