Add objectClass to uit ldap user export
By adding the displayName attribute no matter what, we don't check if the
corresponding objectClass sambaSamAccount has been added for the user.
This makes the LDAP schema sad.

This commit adds a custom objectClass uitAccount for UiT users so that
the attribute does not break anything. This is a temporary fix and the
underlying problem needs to be fixed.