Rewrite suggest_uname for Accounts
Before this commit UiT had it's own special method that took different
arguments than the others. In addition to this the suggest_uname method
of the others took a pointless domain argument that was not used.

The new version of the method is mostly the same as before, but the
domain, fname and lname arguments have been replaced by a populated
person object. The UiT method now has the same as the others instead of
external id, fname, lname. The methods that need fname and lname to
generatre a user name get that from the person object, and UiT gets
external id from the person object as well.

Of course all places where the method was used have been checked and
some lines were removed since they were no longer needed.

Issue: CRB-3015