Change audit dbrecord comparison and sorting
This commit changes the comparison operators for the DbAuditRecord
objects. Comparing the timestamps alone is not enough for sorting
records in an obvious way - because of the time resolution, a lot of
changes will appear to happen at the same time.

This commit changes the sorting of `history show` to sort by timestamp
*and* record_id

Issue: CRB-2560