Check users in other projects for suggest username
When generating usernames for TSD projects the usernames have two parts.
The prefix is the project number, and the suffix is the username. When
checking if a username has been used before it is not enough to check if
the whole name is used, instead we have to check if the suffix has been
used with any prefix.

This commit adds this check to the part of the flow where the user
requests a username. The alternate path already checks this.

Issue: CRB-3017