Add check for SO numbers, don't check the date
It does not make sense to check date for these numbers since we
practicly can't know (based on 2 numbers) when they're born.

Practicly we know by any reasonble knowledge that they're born in either
1900's or 2000's but not which. With normal norwegian FNR's. Løpenummer
that starts a given number are contained for each century. But no such
case with SO numbers. They start at (løpenummer) 10000 and loops through until the
the first unused number with a matching controlschiffer.