Since we actually use FNR to get full birth date???
We can't just return the input, and since S-Number don't really
have a structure to calculate a certain century based on it's rules
we have to make some asumptions.

Current assumptions is, people born current year + 1 (21) is probably
not born in 2021 but in 1921, all the way up to 99.
People born in 00 -> current year is probably not 100 years old
and henceforth is born after year 2000