Fix evu import for the general fs import
For some unknown reason it was assumed that evu entries would only exist
together with other entries. This made the import crash if an evu entry
was alone, or if the previous entry did not contain the key
'studieprogramkode'.

To fix this we now use only the information found in the evu entry
itself. We still check if there is an 'aktiv' entry for the stedkode for
this person already, just in a slightly different way.

Resolves sentry issue: 64652