Refactor hidden bofh cmd access_list_alterable
The access list_alterables command is only used by brukerinfo, and it is
only interested in which groups a user can moderate. This commit changes
the method from checking who has the old opset for group administration
to the checking the group admin and group moderator tables through the
use of the GroupRoles module.

Issue: CRB-2936