Add autoconf for gpg tests
This commit adds a conftest for tests/test_gpg/ that sets up a new,
temporary gpg homedir, gpgme context and a gpg key that can be used
throughout the tests.

The existing tests are updated to use the new key, rather than rely on
one already being pre-configured prior to running the testsuite.

Issue: CRB-1039