Add grace period for affiliation based groups
Persons used to be deleted from ADs affiliation based groups as soon
as the relevant affiliation was deleted. This change adds support for
a grace period so that persons are not deleted from groups until a
certain time after the deletion of the affiliation. Note that all
adconfs using AffGroupSync must be updated correspondingly.

Issue: CRB-3064